+++ src/neural_net/mod.rs
@@ -19,10 +19,10 @@ pub struct NeuralNet{
     pub hidden_layers: usize,
     pub hidden_nodes: usize,
     pub output_nodes: usize,
-    pub learning_rate: f64,
+    pub learning_rate: f32,
     pub activation_function_key: String,
-    pub weights: Vec<Matrix<f64>>,
-    pub biases: Vec<Matrix<f64>>
+    pub weights: Vec<Matrix<f32>>,
+    pub biases: Vec<Matrix<f32>>
 }
 
 impl Eq for NeuralNet {
@@ -37,14 +37,14 @@ impl NeuralNet {
             hidden_layers,
             hidden_nodes,
             output_nodes,
-            learning_rate: 0.1f64,
+            learning_rate: 0.1f32,
             activation_function_key: String::from("SIGMOID"),
             weights: NeuralNet::init_weights(input_nodes, hidden_nodes, output_nodes, hidden_layers),
             biases: NeuralNet::init_biases(hidden_nodes, output_nodes, hidden_layers)
         }
     }
 
-    fn init_weights(input_nodes:usize, hidden_nodes:usize, output_nodes:usize, hidden_layers:usize) -> Vec<Matrix<f64>> {
+    fn init_weights(input_nodes:usize, hidden_nodes:usize, output_nodes:usize, hidden_layers:usize) -> Vec<Matrix<f32>> {
         let mut weights = Vec::new();
     
         for num in 0..hidden_layers+1 {
@@ -60,7 +60,7 @@ impl NeuralNet {
         weights
     }
     
-    fn init_biases(hidden_nodes:usize, output_nodes:usize, hidden_layers:usize) -> Vec<Matrix<f64>> {
+    fn init_biases(hidden_nodes:usize, output_nodes:usize, hidden_layers:usize) -> Vec<Matrix<f32>> {
         let mut biases = Vec::new();
     
         for num in 0..hidden_layers+1 {
@@ -76,7 +76,7 @@ impl NeuralNet {
 }
 
 impl NeuralNet {
-    pub fn guess(&self, input: Vec<f64>) -> Vec<f64> {
+    pub fn guess(&self, input: Vec<f32>) -> Vec<f32> {
         assert!(input.len() == self.input_nodes, "Wrong dimensions!");
         
         let fun: SigmoidActivationFunction = SigmoidActivationFunction{};
@@ -117,13 +117,13 @@ impl NeuralNet {
         }
     }
 
-    fn calculate_layer(weights: &Matrix<f64>, biases: &Matrix<f64>, input: &Matrix<f64>, func: &dyn ActivationFunction) -> Matrix<f64> {
+    fn calculate_layer(weights: &Matrix<f32>, biases: &Matrix<f32>, input: &Matrix<f32>, func: &dyn ActivationFunction) -> Matrix<f32> {
         let result = weights.mul(input) + biases;
 
         NeuralNet::apply_activation_function(result, false, func)
     }
 
-    fn apply_activation_function(input: Matrix<f64>, derivative: bool, func: &dyn ActivationFunction) -> Matrix<f64> {
+    fn apply_activation_function(input: Matrix<f32>, derivative: bool, func: &dyn ActivationFunction) -> Matrix<f32> {
         if derivative {
             func.apply_derivative_of_activation_func_to_matrix(input)
         } else {
